Rewrite volume handling, one side effect is that options are now supported
Before this, the driver had no notion of volumes it only kept track of an array
of names which happened to be path/mount point. Now we keep track of a all
mounts as a "beegfsMount". This allows for keeping track of names and paths
separatly. As a bonus, we can now create volumes on non-default beegfs mounts.
i.e. "docker volume create -d beegfs --name b3 -o root=/stor/b3"